BZ 963208 - NullPointerException when trying to delete duplicated workspace

+++ modules/dashboard-ui/dashboard-ui-panels/src/main/java/org/jboss/dashboard/ui/panel/advancedHTML/HTMLDriver.java
@@ -80,7 +80,7 @@ public class HTMLDriver extends PanelDriver implements Exportable {
         new HibernateTxFragment() {
         protected void txFragment(Session session) throws Exception {
             HTMLText htmlText = load(instance);
-            htmlText.delete();
+            if (htmlText != null) htmlText.delete();
         }}.execute();
     }
